+++ _tools/run/helpers/lib/refreshIndexes.js
@@ -12,22 +12,20 @@ const configsObject = require('./configsObject.js')
 // Refresh indexes
 async function refreshIndexes (argv) {
   try {
-    await fs.emptyDir(process.cwd() + '/_site')
     /*
     Much of _tools doesn't properly cater for the baseurl.
     The assumption in most methods is that baseurl is ''.
     Setting baseurl here is a temporary/safe fix to avoid further regressions for now.
-    TODO: rewrite all of _tools to address the following issues:
-      – Functions too tightly coupled across disparate operations
-      - Lacks clear order of operations
-      – The 'helpers' are not pure functions (i.e. they have many, hard-to-trace side effects)
     */
     const _argv = {
       ...argv,
       baseurl: ''
     }
-    await webpack(_argv)
-    await jekyll(_argv)
+    if (!_argv.skipbuild) {
+      await fs.emptyDir(process.cwd() + '/_site')
+      await webpack(_argv)
+      await jekyll(_argv)
+    }
     if (argv.format === 'print-pdf' ||
       argv.format === 'screen-pdf' ||
       argv.format === 'epub') {
